backend/requirements.txt
Flask==2.2.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
flask-apispec==0.11.4
gunicorn==21.2.0
Werkzeug==2.2.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.23
PyMySQL==1.1.0
alembic==1.12.1
marshmallow==3.20.1
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.200
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.25.2
scipy==1.11.3
python-dotenv==1.0.0
pydantic==2.4.2
email-validator==2.1.0
requests==2.31.0
python-magic==0.4.27
passlib==1.7.4
bcrypt==4.0.1
pytest==7.4.3 